exception tava errada
