+++ aplicacao-grafica-samu/src/main/java/br/ufrrj/samu/exceptions/AlreadyExistsException.java
@@ -1,0 +1,15 @@
+package br.ufrrj.samu.exceptions;
+
+public class AlreadyExistsException extends Exception {
+
+    public AlreadyExistsException() {
+    }
+
+    public AlreadyExistsException(String message) {
+        super(message);
+    }
+
+    public AlreadyExistsException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
