+++ aplicacao-grafica-samu/src/main/java/br/ufrrj/samu/exceptions/UnknownUserException.java
@@ -16,17 +16,4 @@ public class UnknownUserException extends Exception {
         super(cause);
     }
 
-    public static class AlreadyExistsException extends Exception {
-
-        public AlreadyExistsException() {
-        }
-
-        public AlreadyExistsException(String message) {
-            super(message);
-        }
-
-        public AlreadyExistsException(String message, Throwable cause) {
-            super(message, cause);
-        }
-    }
 }
